+++ limbo/testcases/webpki/__init__.py
@@ -12,6 +12,8 @@ from limbo.assets import ASSETS_PATH, Certificate, ext
 from limbo.models import Feature, KeyUsage, KnownEKUs, PeerName
 from limbo.testcases._core import Builder, testcase
 
+from .aki import *  # noqa: F403
+
 
 @testcase
 def cryptographydotio_chain(builder: Builder) -> None:
@@ -461,110 +463,6 @@ def root_with_extkeyusage(builder: Builder) -> None:
         .peer_certificate(leaf)
         .fails()
     )
-
-
-@testcase
-def root_with_aki_authoritycertissuer(builder: Builder) -> None:
-    """
-    Produces the following **invalid** chain:
-
-    ```
-    root -> EE
-    ```
-
-    The root cert includes the authorityKeyIdentifier extension with the
-    authorityCertIssuer field, which is forbidden under the [CA/B BR profile]:
-
-    > 7.1.2.1.3 Root CA Authority Key Identifier
-    > Field                 Description
-    > ...
-    > authorityCertIssuer   MUST NOT be present
-
-    [CA/B BR profile]: https://cabforum.org/wp-content/uploads/CA-Browser-Forum-BR-v2.0.0.pdf
-    """
-
-    key = ec.generate_private_key(ec.SECP256R1())
-    dirname = x509.DirectoryName(x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, "myCN")]))
-    aki = x509.AuthorityKeyIdentifier.from_issuer_public_key(key.public_key())
-    # Manually add the authorityCertIssuer field, since cryptography's API doesn't allow to
-    # add it without also specifying the authorityCertSerialNumber field
-    aki._authority_cert_issuer = [dirname]
-
-    root = builder.root_ca(key=key, aki=ext(aki, critical=False))
-    leaf = builder.leaf_cert(root)
-
-    builder = builder.server_validation()
-    builder = builder.trusted_certs(root).peer_certificate(leaf).fails()
-
-
-@testcase
-def root_with_aki_authoritycertserialnumber(builder: Builder) -> None:
-    """
-    Produces the following **invalid** chain:
-
-    ```
-    root -> EE
-    ```
-
-    The root cert includes the authorityKeyIdentifier extension with the
-    authorityCertSerialNumber field, which is forbidden under the
-    [CA/B BR profile]:
-
-    > 7.1.2.1.3 Root CA Authority Key Identifier
-    > Field                         Description
-    > ...
-    > authorityCertSerialNumber     MUST NOT be present
-
-    [CA/B BR profile]: https://cabforum.org/wp-content/uploads/CA-Browser-Forum-BR-v2.0.0.pdf
-    """
-
-    key = ec.generate_private_key(ec.SECP256R1())
-    aki = x509.AuthorityKeyIdentifier.from_issuer_public_key(key.public_key())
-    # Manually add the authorityCertSerialNumber field, since cryptography's
-    # API doesn't allow to add it without also specifying the
-    # authorityCertIssuer field
-    aki._authority_cert_serial_number = 1234
-
-    root = builder.root_ca(key=key, aki=ext(aki, critical=False))
-    leaf = builder.leaf_cert(root)
-
-    builder = builder.server_validation()
-    builder = builder.trusted_certs(root).peer_certificate(leaf).fails()
-
-
-@testcase
-def root_with_aki_all_fields(builder: Builder) -> None:
-    """
-    Produces the following **invalid** chain:
-
-    ```
-    root -> EE
-    ```
-
-    The root cert includes the authorityKeyIdentifier extension with the
-    authorityCertIssuer and authorityCertSerialNumber fields, which is
-    forbidden under the [CA/B BR profile]:
-
-    > 7.1.2.1.3 Root CA Authority Key Identifier
-    > Field                         Description
-    > ...
-    > authorityCertIssuer           MUST NOT be present
-    > authorityCertSerialNumber     MUST NOT be present
-
-    [CA/B BR profile]: https://cabforum.org/wp-content/uploads/CA-Browser-Forum-BR-v2.0.0.pdf
-    """
-
-    key = ec.generate_private_key(ec.SECP256R1())
-    key_identifier = x509.AuthorityKeyIdentifier.from_issuer_public_key(
-        key.public_key()
-    ).key_identifier
-    dirname = x509.DirectoryName(x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, "myCN")]))
-    aki = x509.AuthorityKeyIdentifier(key_identifier, [dirname], 1234)
-    root = builder.root_ca(key=key, aki=ext(aki, critical=False))
-    leaf = builder.leaf_cert(root)
-
-    builder = builder.server_validation()
-    builder = builder.trusted_certs(root).peer_certificate(leaf).fails()
 
 
 @testcase
